Extract PairLocationInput, reset mapbox token
