+++ src/frontend/boilerplates/archipelago/.env
@@ -1,2 +1,3 @@
 REACT_APP_MIDDLEWARE_URL=http://localhost:3000/
+REACT_APP_MAPBOX_ACCESS_TOKEN=
 PORT=5000
