+++ src/frontend/boilerplates/archipelago/src/resources/Agent/Activity/Event/EventEdit.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { SimpleForm, TextInput, DateTimeInput } from 'react-admin';
 import MarkdownInput from 'ra-input-markdown';
 import { Edit } from '@semapps/archipelago-layout';
-import { ActorsInput, ThemesInput } from '../../../../inputs';
+import { ActorsInput, ThemesInput } from '../../../../pair';
 import EventTitle from './EventTitle';
 
 const EventEdit = props => (
