+++ src/frontend/boilerplates/archipelago/src/resources/Resource/Skill/SkillEdit.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { SimpleForm, TextInput } from 'react-admin';
 import { Edit } from '@semapps/archipelago-layout';
-import { UsersInput } from '../../../inputs';
+import { UsersInput } from '../../../pair';
 import SkillTitle from './SkillTitle';
 
 export const SkillEdit = props => (
